the updated visual representation of water * small bug fixes in the code

+++ GeneralTrainingEnvironmentForMAS/Assets/Examples/AntColony/Scripts/AntColonyFitness.cs
@@ -7,8 +7,8 @@ using System.Threading.Tasks;
 public class AntColonyFitness : Fitness
 {
 
-    public static Dictionary<string, int> FitnessValues = new Dictionary<string, int> {
-        { "AgentBringFood", 1 },
+    public static Dictionary<string, float> FitnessValues = new Dictionary<string, float> {
+        { "AgentBringFood", 1},
     };
 
     public static string[] Keys = FitnessValues.Keys.ToArray();
